source/requirements.txt
azure-storage-queue
azure-storage-blob
azure-identity
python-dotenv
